Move screen and scroll logic to default.vue
